Major revamp of Body implementation.
Before, the Body implementation was:

- BodyRaw holds a pointer and some other data, and implements all the
  bindings to C functions.
- Body holds a BodyRaw, and forwards many function calls to it.
  The Body can be duplicated to create multiple non-thread-safe
  references.

Now, the Body implementation is:

- Body holds a pointer and implements the bindings to C functions.
- BodyHandle (i.e. Handle<Body>) holds a Body. The BodyHandle can be
  cloned to create multiple thread-safe references.

This implementation is considerably simpler, because there is no need
to forward many function calls. Also, there is no need for
PhantomData or a type parameter.

Another advantage is that, in theory, it should be safe to access the
Body from multiple threads. This would allow e.g. the physics engine
to run in a separate thread from the rendering code, which reads the
body positions and angles. Of course, the body should be write-locked
while the Space is being stepped. (That is not yet implemented.)

The new Body implementation does not yet support user data.
A new implementation of user data will be added in the future.