Added CpVect struct.
It is a duplicate of the cpVect type from chipmunk-sys, but with
extra methods equivalent to all the "cpv*" functions from Chipmunk.
In the future, functions in this crate will accept and return CpVect
instances instead of (f64, f64).

CpVect is not intended as a full replacement for crates like cgmath
and nalgebra, but it will allow users to perform basic 2D vector math
without converting back and forth between types.