Optional cgmath and nalgebra integration.
This crate can be compiled with the "cgmath" and/or "nalgebra"
features to enable support for converting CpVect to and from
cgmath::Vector2<f64>, nalgebra::Point2<f64>, and
nalgebra::Vector2<f64>.

These features are not enabled by default because they add cgmath
and/or nalgebra as dependencies, and most users will be using at most
one of those crates, not both.

This support needs to be built into this crate (or else those crates)
because user code cannot implement traits involving only types from
extern crates.