Handle is no longer thread-safe.
Handle is now based on Rc and RefCell, and WeakHandle is based on
Weak (from std::rc, not std::sync) and RefCell.

I'm abandoning the goal of thread-safety for now. There are too many
complications and risks (e.g. deadlocks), and using Arc/Mutex adds
significant performance costs, even in single-threaded programs.

It may be possible to achieve thread safety in the future, but it
should be opt-in, using different types, so that you only pay the
cost if you need it.